+++ examples/testbed_simple/src/BranchMethods.js
@@ -173,7 +173,8 @@ class BranchMethods extends Component {
           customData: {
             "Custom_Event_Property_Key1": "Custom_Event_Property_val1",
             "Custom_Event_Property_Key2": "Custom_Event_Property_val2"
-          }
+          },
+          alias: 'ItemViewed'
         }
       )
       branchEvent.logEvent()
@@ -213,6 +214,29 @@ class BranchMethods extends Component {
     } catch (err) {
       console.log('sendStandardEvent err', err)
       this.addResult('error', 'sendStandardEvent', err.toString())
+    }
+  }
+
+  openURL = async () => {
+    const url = 'https://zjef.test-app.link/n4efBZnSu8';
+    try {
+      await branch.openURL(url);
+      this.addResult('success', 'openURL', url);
+    }
+    catch (err) {
+      this.addResult('error', 'openURL', err.toString());
+    }
+  }
+
+  lastAttributedTouchData = async() => {
+    const attributionWindow = 365
+    try {
+      let latd = await branch.lastAttributedTouchData(attributionWindow)
+      console.log('lastAttributedTouchData', latd)
+      this.addResult('success', 'lastAttributedTouchData', latd)
+    } catch (err) {
+      console.log('lastAttributedTouchData', err)
+      this.addResult('error', 'lastAttributedTouchData', err.toString())
     }
   }
 
@@ -256,6 +280,8 @@ class BranchMethods extends Component {
           <Button onPress={this.getCreditHistory}>getCreditHistory</Button>
           <Button onPress={this.logStandardEvent}>BranchEvent.logEvent (Standard)</Button>
           <Button onPress={this.logCustomEvent}>BranchEvent.logEvent (Custom)</Button>
+          <Button onPress={this.openURL}>openURL</Button>
+          <Button onPress={this.lastAttributedTouchData}>lastAttributedTouchData</Button>
         </ScrollView>
       </View>
     )